Temporarily exclude linux-vdo ISCSIDevice tests.
There's an incompatibility with the population of
/sys/kernel/config/target/iscsi and what targetcli expects to find
there.  As a result  targetcli attempts to run os.path.listdir
against a regular file which throws an exception.  This prevents the
cleanup of the test created devices and results in leaking farms.
This occurs on both Fedora 35 and 36.

See VDO-5437.

Pair: jwiele

.../vdotests.suites
Added ISCSIDevice tests to exclusion list.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156347]