Update journal state after save and resume. [VDO-5417, BZ 2064802]
The "save" code path saves all dirty metadata blocks in each
section of the code in phases but bypasses some block-by-block
in-memory data structure updates (which wouldn't matter if the
device gets shut down next), and resets some state on resume
if needed, rather than trying to negotiate the simultaneous
save and cleanup of multiple interrelated journals and other
data structures.

However, the resets done are incomplete, and can leave some
journal pages unavailable for use (which may not be clear
until it comes time to write those pages), and an incorrect
notion of how many slots in the recovery journal are available
to be filled.

Pair: corwin,msakai

.../slab-journal.h
.../slab-journal.c
.../block-allocator.c
After a successful suspend with saving and resume, reopen the
  slab journal to reset the state.

.../recovery-journal.c
Reset available_space in initialize_journal_state, so it gets
  done during a resume after save as well.

.../vdoTestBase.h
.../vdoTestBase.c
.../GrowLogical_t1.c
Add tests for doing grow-logical operations with saving at
  suspend time as well as without saving.

.../SuspendResume_t2.c
New test. Test with enough I/O after suspend-resume to cycle
  through all blocks of the recovery journal and some slab
  journals (a few times over) to ensure that none of the
  blocks were left locked in the suspend-resume process. Test
  multiple suspend-resume cycles to exaggerate any resource
  loss like available_space not getting reset.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156411]