Made InFlightDedupeAndCompress not use /permabit/datasets [VDO-5419]
The test was written to use /permabit/datasets/vdo/iobw.tst, a
data file generated by Iometer that has sequential repetition of
blocks suitable for concurrent dedupe. That file won't be
available to OSS contributors, so to make this test usable for
them, this change dynamically generates a data file to use instead
of the Iometer dataset.

Pair: raeburn, msakai

.../InFlightDedupeAndCompress.pm
Used generateDataFile() to create unique input data, and then
  piped that data to a small in-line perl script that repeats
  each block six times (like iobw.tst does). Used the output
  of that script to replace the /permabit/datasets/ file.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156474]