Removed the work_queue field from dedupe_index [VDO-5405].
This small clean-up is a prequel to actually fixing the bug. With the
recent changes to dedupe-index.c, the queue is only referenced directly
when shutting down, but finish_vdo() already finishes all of the
queues, and vdo_destroy() frees them.

Pair: msakai

.../dedupe-index.c
Removed the 'uds_queue' field from struct dedupe_index, the finishing
  of it in vdo_finish_dedupe_index(), and the freeing of it in
  vdo_free_dedupe_index().

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156521]