Rename hash-zone.[hc] to dedupe.[hc] preparatory to fixing [VDO-5405].
This is the first of a series of changes which fixes the race in
handling dedupe timeouts. These changes will also separate
dedupe_contexts from data_vios. Doing so simplifies both the handling
of index queries and the reuse of data_vios. Finally, the management of
dedupe_contexts will be done by the hash zones, rather than the
dedupe_thread (dedupe-index). While doing so requires a pool of
contexts per hash zone, it both simplifies making and timing out index
queries, and eliminates several thread transitions. A general outline
of the series of ensuing changes follows:

1) Code motion: hash-zone.[hc], hash-lock.[hc], and dedupe-index.[hc]
will be combined into a single pair of files: dedupe.[hc]. Once dedupe
queries are mediated by the hash zones rather than the dedupe index,
the dedupe thread's only purpose will be to perform administrative
actions on the dedupe index. Those administrative actions require
careful sequencing with the draining of the hash zones, and combining
them under a single action_manager makes this simple. It also has the
side-benefit of being a step towards meeting the requirement that we
have fewer source files. It is also the case that once lvm ceases to
use dm-setup messages for managing the index, and we move the test-only
index filler out of the kernel, most of the code now in dedupe-index.c
will go away.

2) Structure merging: struct dedupe_index and struct hash_zones will be
combined into a single structure (name TBD). An action manager will be
introduced to manage suspend and resume of both the index and the
hash zones.

3) The bug fix itself:
- Creating a pool of dedupe_contexts in each hash zone.
- Running a timer in each zone for timing out index queries.
- Converting the timer handler to only enqueue a completion to run on
  the hash zone.
- Converting the uds callback to only enqueue the requesting data_vio
  to run on the hash zone (unless the request has already timed out).
All of the management of contexts and the lists on which they live will
be done only on the hash zone thread, thereby eliminating the race.

Pair: msakai

.../hash-zone.h
.../hash-zone.c
.../dedupe.h
.../dedupe.c
.../Makefile.common
.../MANIFEST.yaml
Renamed hash-zone.[hc] to dedupe.[hc].

.../data-vio.h
.../hash-lock.c
.../vdo-load.c
.../vdo.c
.../HashZone_t1.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156524]