Implement userspace dm-bufio functions.
This change creates a fake partial implementation of dm-bufio in
userspace, in order to enable us to remove all the #ifdefs that are
currently used to switch between I/O systems. This change also
obsoletes a number of userspace utilies which are being replaced by
the new fake dm-bufio.

The fake dm_bufio client makes the backing file and the buffers. The
fake dm-buffer is a wrapper around a page-sized byte buffer, and the
fake block_device structure is a wrapper around a file descriptor.
In order to try to retain some userspace efficiencies, a released
dm_buffer does not free its memory, but goes on a list to be
repurposed for future buffer requests. The client then frees all
these buffers when destroyed.

The kernel side of the code is not affected by this change.

Pair:  jwiele

.../dm-bufio.h
.../dm-bufio.c
.../uds/userLinux/uds/linux/blkdev.h
.../blkdev.c
Added fake implementations on dm-bufio and block_device for userspace,
  base on the old io_region code.

.../vdo/fake/linux/blkdev.h
.../uds/userLinux/uds/linux/blkdev.h
Moved the existing VDO fake blkdev.h to UDS.

.../io-factory.h
.../src/uds/io-factory.c
.../index-layout.h
.../index-layout.c
.../volume-store.h
.../volume-store.c
.../buffered-reader.h
.../buffered-reader.c
.../buffered-writer.h
.../buffered-writer.c
Removed userspace IO functions, using the kernel-side code for
  both builds.

.../kernelLinux/uds/io-factory.c
.../src/uds/io-factory.c
.../kernelLinux/tests/Bufio_t1.c
.../src/tests/Bufio_t1.c
Moved the kernel io-factory implementation into src/uds so it can
  be used by both builds, along with its test.

.../ioFactory.c
.../ioRegion.h
.../fileIORegion.h
.../fileIORegion.c
Removed unused userspace ioFactory implementation, as well as
  ioRegion and fileIORegion, which now have no users.

.../buffered-reader.h
.../buffered-reader.c
Changed buffered_reader to always return UDS_OUT_OF_RANGE rather
  that UDS_SHORT_READ if any part of the read request is out of
  range. UDS only ever reads full pages so in practice this makes
  no difference.

.../defines
.../c++/uds/kernelLinux/build/Makefile
.../c++/uds/userLinux/build/Makefile
.../MANIFEST.yaml
.../packaging/src-dist/user/utils/uds/Makefile
Adjusted manifests and Makefiles for deleted and moved files.

.../compiler.h
.../assertions.h
Added handling for kernel-style errors to userspace now that some
  functions can return them.

.../uds/userLinux/uds/linux/blkdev.h
Moved SECTOR_SIZE and SECTOR_SHIFT to blkdev.h to mirror where the
  kernel defines them.

.../buffered-reader.h
.../type-defs.h
Moved userspace definition of sector_t to type-defs.h where it makes
  more sense.

.../errors.h
.../errors.c
.../Volume_t3.c
Removed UDS_BAD_IO_DIRECTION and UDS_END_OF_FILE which are no
  longer used by anything.
Replaced existing test uses of UDS_SHORT_READ with UDS_OUT_OF_RANGE
  and removed UDS_SHORT_READ, which is otherwise unused.

.../convertToLVM.c
Rewrote conversion to work with the new fake dm-bufio. The
  previous chapter moving code made assumptions about the page
  buffers which are not true with the dm-bufio interface.

.../albtest.c
Ensured that test index files exist before running tests. This is
  necessary because the block_device interface has no concept of
  "creating" a block device, so tests must pre-create any files they
  might need.

.../copyDevice.c
.../getTestIOFactory.c
.../BufferedReader_t1.c
.../ConvertToLVM_n1.c
.../ConvertToLVM_x1.c
.../StatusCodes_t1.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156554]