Copy hash-lock.[hc] into dedupe.[hc] [VDO-5404].
This is the third step in the code motion portion of the fix layed out
in change 156524. It copys the contents of the hash lock files directly
into the dedupe files. The next change will de-externalize the portions
of dedupe.[hc] which no longer have external callers.

Pair: sclafani

.../hash-lock.h
.../hash-lock.c
.../dedupe.h
.../dedupe.c
.../Makefile.common
.../MANIFEST.yaml
Copied hash-lock.[hc] into dedupe.[hc] and deleted the former.

.../packer.c
.../physical-zone.c
.../vio-write.c
.../Packer_t1.c
.../VDO_t4.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156558]