Use bits_per() instead of compute_bits().
The kernel library provides bits_per() to calculate the number
of bits required to store an item of a particular size. There
is one difference between compute_bits() and bits_per(), which
is the value returned if the input is zero. UDS never calls
compute_bits with a zero value, so this difference is
irrelevant.

Pair: joe, msakai

.../log2.h
Add a user space header with the same name as the kernel
  header that provides bits_per().

.../hash-utils.h
.../hash-utils.c
.../defines
.../uds-module.c
.../MANIFEST.yaml
.../Makefile
Remove compute_bits() and a not very useful static assertion,
  then remove hash-utils.c since those were the only things in
  hash-utils.c.

.../delta-index.c
.../geometry.c
.../open-chapter.c
.../volume-index005.c
Use bits_per() instead of compute_bits().

.../HashUtils_t1.c
Delete because the test of compute_bits was the last test in
  the suite.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156575]