Copy dedupe-index.[hc] into dedupe.[hc] [VDO-5404].
This is the sixth step in the code motion portion of the fix laid out
in change 156524. It merges dedupe-index into dedupe, paving the way
for the merging of struct dedupe_index and struct hash_zones.

Pair: sclafani

.../dedupe-index.h
.../dedupe-index.c
.../dedupe.h
.../dedupe.c
.../Makefile.common
.../MANIFEST.yaml
Copied dedupe-index.[hc] into dedupe.[hc].

.../dm-vdo-target.c
.../dump.c
.../pool-sysfs.c
.../sysfs.c
.../vdo-load.c
.../vdo-resume.c
.../vdo-suspend.c
.../vdo.c
.../asyncLayer.c
.../CMessage.pm
.../CSysfs.pm
Converted all includes of dedupe-index.h to dedupe.h.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156581]