Create a proper fake linux/completion.h.
The comments in the fake linux/delay.h claim that we get completion.h
in the kernel by including delay.h, but this seems to not be correct.
Instead, the code somehow picks up completion.h indirectly as a result
of including logger.h from UDS (which itself is indirectly included
through permassert.h from UDS). In order to be less confusing, this
change moves the fake struct completion declarations to a new fake
linux/completion.h so that the kernel includes can be more direct.

Pair:  corwin

.../completion.h
.../delay.h
Moved struct completion declarations from delay.h to completion.h.

.../admin-completion.h
.../admin-completion.c
.../sync-completion.c
Added include of linux/completion.h.

.../mutexUtils.c
Adjusted comments on the struct completion implementation for tests.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156621]