Fix index suspend and resume hangs [VDO-5458] and [VDO-5459].
When the dedupe_index suspended flag was replaced with an admin_state
in hash_zones as part of [VDO-5405], the handling of the state on
suspend was incorrect. When the index was suspended, the state was not
transitioned from a quiescing to a quiescent state, which prevented
resume_index() from ever resuming. Ultimately, this lead to the index
never coming back on line, and somehow also caused lvresize to fail.

Pair: sclafani

.../dedupe.c
Modified suspend_index() to properly transition the state on suspend.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156636]