Rename the state_lock field of struct hash_zones to lock [VDO-5405].
This is the first change in the third (actual bug fix) portion of the
dedupe index fix laid out in change 156524. Currently, there are two
locks in struct hash_zones: state_lock and pending_lock. The latter
will be going away, at which point it makes no sense to distinguish
them. This change merely renames the state_lock field to 'lock' in
order to minimize the diff of the actual fix.

Pair: raeburn

.../dedupe.c
Renamed the 'state_lock' field of struct hash_zones to 'lock'.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156645]