Divorce dedupe contexts from data vios and fix [VDO-5405].
This is the final change in the third portion of the dedupe index fix
which actually fixes the bug.

This change divorces dedupe contexts from data vios, adding a pool of
dedupe contexts in each hash zone. When a hash lock wishes to make a
dedupe query on behalf of a data_vio, it attempts to acquire a context
from its pool. If none are available, the index query is skipped. There
are a number of benefits to this respelling. Index queries no longer
need to transition to and from the dedupe thread, which means there is
no need to check whether that thread is blocked rebuilding the
index. If the index is rebuilding the index query will fail
immediately. This eliminates both some thread transitions and some
locking.

The fix also entails a complete respelling of the handling of both
dedupe callbacks and timeouts. For a query which has not timed out, the
dedupe callback merely continues the data_vio making the query to run
in its hash zone. For a query which has timed out, the dedupe callback
merely modifies the state of the context to indicate that it is both
complete and timed out. For timeouts, the timer callback merely
enqueues a new hash_zone completion to run on the hash zone thread. The
callback of that complete processes the timeout. These changes mean
that all the handling and state changes of data_vios, dedupe_contexts,
and the lists which hold the latter is done on the appropriate hash
zone thread. Doing so eliminates the races in the original bug.

Pair: msakai

.../dedupe.h
.../dedupe.c
Reworked the mechanisms for index queries:
- Added a pool of dedupe contexts to each hash zone which not only
  includes the contexts themselves, but a timer, the completion for
  sending timeout notifications, some state management fields, and
  statistics.
- Removed the pending_lock and all of the dedupe context management
  fields (originally in struct dedupe_index) from struct hash_zones.
- Pulled struct periodic_event_reporter into struct hash_zones and
  reduced it to a rate limiter, an atomic timeouts counter, and a
  counter of timeouts reported.
- Named the (previously anonymous) enum representing dedupe context
  states 'dedupe_context_state', and expanded the number of states from
  3 to 5 (see the doc in dedupe.c for an explanation of the states).
- Renamed vdo_query_index() to query_index() and unexposed it. Modified
  it to attempt to acquire a dedupe context, continuing the data_vio
  without deduplicating if there are none available.
- Modified start_updating() and start_querying() to call query_index()
  nee vdo_query_index() directly since this call should now be done in
  the hash_zone not on the dedupe thread. This eliminates the need to
  call data_vio_may_query_index() (which has been removed along with
  the 'deduping' flag from struct hash_zones, which it consulted) as
  the queries now can't get stuck waiting on the dedupe thread if the
  index is rebuilding (and if the index is offline, the queries will
  return with an error immediately). The callbacks used to launch the
  query from the dedupe thread have also been removed (update_index()
  and query_index()).
- Expanded check_for_drain_complete() to check that there are no
  outstanding dedupe queries and to ensure that the timer won't fire
  and that the timer notification is not in flight.
- Pulled the actual processing of query results out of
  finish_index_operation() (the dedupe callback) and into
  finish_updating() and finish_querying() both of which run in the
  hash_zone. This new function also returns the dedupe context to the
  available list. Modified finish_index_operation() itself to merely
  update the state of the context, and, if not timed out, continue the
  context's data_vio.
- Pulled get_dedupe_index_timeout() into start_expiration_timer(),
  changed the latter to take only a dedupe_context, and eliminated
  start_expiration_timer_for_vio(). Timers are now always set in
  reference to some dedupe context, so this simpler interface is all
  that is required.
- Simplified report_dedupe_timeouts() to manage the reporting and rate
  limiting directly as it is now called in hash_zones, not in the timer
  callback (which runs at interrupt context). All of the other
  reporting methods have been removed. The timer and event report
  cleanup has been removed from finish_dedupe_index().
Renamed vdo_get_hash_zone_statistics() to vdo_get_dedupe_statistics()
  and combined it with vdo_get_dedupe_index_statistics(). This new
  single method tabulates all index and hash zone related statistics
  and fills in a vdo_statistics structure directly. The iteration over
  the zones is now handled in this method rather than in vdo.c. This
  is a more convenient spelling now that the per-hash_zone pools of
  dedupe contexts require tabulating.
Factored dump_hash_zone() out of vdo_dump_hash_zones().
Added a unit test hook for index queries.

.../data-vio.h
.../dedupe.c
Moved the definition of struct dedupe_context from data-vio.h into
  dedupe.c and restructured it to reflect the new handling of index
  queries.

.../data-vio.h
Changed the dedupe_context field to be a struct dedupe_context * rather
  than a struct dedupe_context, and moved the field into the portion of
  the data_vio which is reset by the data_vio_pool.

.../VDOStatistics.stats
Moved the currDedupeQueries stat into HashLockStatistics as this is
  now tracked by the hash zones.
Removed the maxDedupeQueries stat from IndexStatistics. Since each hash
  zone has its own pool of queries, and there can be timed out queries
  in each zone, it is no longer clear what this stat means or how to
  compute it.
Bumped version.

.../atomic-stats.h
.../dedupe.c
Moved the dedupe_context_busy counter into struct hash_zones now that
  it is the hash zones which manage the dedupe contexts.

.../completion.h
.../completion.c
Added VDO_HASH_ZONE_COMPLETION.

.../data-vio-pool.c
.../vdo.c
Reflected changes.

.../jiffies.h
.../testTimer.h
.../testTimer.c
Converted jiffies to be a call to getUnitTestJiffies() which updates
  its counter each time is called, rather than being based on actual
  clock time. This is useful for the new DedupeTimeouts_t1.

.../timer.h
.../testTimer.h
.../testTimer.c
Added a functional fake of the timer interface to facilitate unit
  testing of dedupe timeouts.

.../Makefile
Added generation of dedupeContext.h for use by new DedupeTimeouts_t1.
It is also needed by HashZone_t1.

.../DedupeTimeouts_t1.c
Created new unit test of dedupe timeouts.

.../DataVIOPool_t1.c
Removed the usleep() since jiffies will now increment each time it is
  called.

.../HashZone_t1.c
.../VDO_t4.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156651]