Fix NULL dereference when no dedupe contexts are available [VDO-5461].
The rework of the handling of index queries in order to fix [VDO-5405]
introduced an NPE in the event that an index query can not be made due
to a lack of available dedupe contexts.

Pair: msakai

.../dedupe.c
Modified finish_updating() and finish_querying() to not attempt to
process the dedupe context if the agent didn't get one.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156677]