Limit the size of file copying in conversion tests.
This fixes a problem resulting from the userspace dm-bufio
implementation. That change made it so userspace test will always
reuse an existing index file. The particular problem is that some
tests, notable Configuration_n1, can make a very large (but sparse)
index file, around 190G. That is not a problem because the test
actually writes very little of that file.

However, ConvertToLVM_n1 and ConvertToLVM_x1 include slide_file(), a
method that copies the contents of the index file. By re-writing the
entire contents of the index file, the method ends up filling in all
the sparse holes in the index file. This is fine of the index file
is only as large as the index used by that test, but if a previous
test has made the index file larger than that, it will attempt to
write up to 190G to disk, which is more storage than most of our
test machines have.

The result is that the ConvertToLVM_* tests fail, and they also causes
any subsequent tests to fail if those tests attempt to write new files.

Pair: sclafani

.../ConvertToLVM_n1.c
.../ConvertToLVM_x1.c
Limit slide_file() to copy just the number of bytes required for
  the test index and not more. This mirrors the behavior of Suspend_n2,
  which already handles a similar circumstance correctly.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156680]