Fix hang in DedupeTimeouts_t1 [VDO-5462].
This new unit test was introduced as part of the fix for [VDO-5405]. It
made the incorrect assumption that the artificial method to fire the
timers was deterministic. In fact it is not, and due to races with the
index, may sometimes result in not sending the message to process
dedupe timeouts, which ultimately results in at least one dedupe
context stuck in the pending state when the test goes to wait for the
write to complete.

Pair: sclafani

.../DedupeTimeouts_t1.c
Added an extra callback and signaled state on which the test will wait
  if fireTimers() actually generates a timeout. When this happens, the
  test thread will continue after the timeout processing has completed.

.../testTimer.h
.../testTimer.c
Modified fireTimers() to return a boolean indicating whether or not
  any timer fired.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156689]