Improve handling of timed-out dedupe contexts.
The fix for VDO-5504 involved a complete rewrite of the handling of
dedupe contexts. This change improves the efficiency of recycling
dedupe contexts whose queries have timed out.

Rather than moving timed out contexts to a timed_out list and then
searching that list for contexts which are now also complete, timed out
contexts are simply removed from the pending list. In the UDS callback
for a timed out context, in addition to transitioning the context to
the timed out complete state, the context is enqueued on a funnel
queue. When a context is required for a new request, but the available
list is empty, rather than searching the timed_out list for a complete
context, the funnel queue is polled once.

Pair: sclafani

.../dedupe.c
Replaced the timed_out list in struct hash_zone with a funnel queue
  named timed_out complete. Added a funnel_queue_entry to struct
  dedupe_context.
Modified timeout_index_operations_callback() to merely remove timed out
  contexts from the pending list, but not place them on the (now
  defunct) timed_out list.
Modified finish_index_operation() to enqueue timed out contexts on the
  timed_out_complete funnel queue.
Modified acquire_context() and check_for_drain_complete() to poll the
  funnel queue rather than searching the timed_out list.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156749]