Modified vdo load to grow to the logical size from the table [VDO-5457].
Change 152087 introduced validation that the logical size specified on
the table line matches the logical size recorded in the super
block. Unfortunately, this checking can cause attempts to create an lvm
managed vdo with default logical size to fail due to the fact that lvm
rounds the logical size to the nearest extent, but vdoformat does not
(BZ 2114006). As a stop-gap, this check has been relaxed, but
unfortunately, while doing so allows default volumes to be created and
started, it exposes other potential problems (BZ 2120738).

When lvm rounds up, clients writing to the logical volume will get
out-of-range errors if they attempt to write to the rounded-up
area. Conversely, for vdo volumes which were converted to lvm from vdo
manager, rounding down would result in a loss of access to any data
previously written to the rounded-down area.

This change addresses the case where lvm rounds up.

Pair: msakai

.../device-config.c
Modified vdo_validate_new_device_config() to not ever reject tables
  which would result in logical growth (previously, these would be
  rejected if the may_grow flag was false, but now that flag only
  pertains to physical growth). Since the advent of the block map tree,
  there is no harm in growing the logical size at initial load time,
  even if the vdo is dirty.

.../vdo-load.c
Modified decode_from_super_block() to accept the logical size specified
  in the device config rather than the super block if the former is
  larger.

.../asyncLayer.c
Modified makeDeviceConfig() to not copy the logical block size from the
  test configuration's vdo_config to the device_config being
  constructed. With the exception of FormatVDO_t2, this copying is
  unnecessary, and doing so gets in the way of unit testing
  auto-growth.

.../vdoTestBase.c
Modified reloadVDO() to update the logical block size recorded in the
  test configuration's vdo_config now that it and the device config are
  not kept in sync by makeDeviceConfig().
Modified modifyVDO() to update the device config to reflect the desired
  growth now that the device config is not automatically kept in sync
  with the test configuration's vdo_config.

.../FormatVDO_t2.c
Modified zeroLogicalBlocks() to zero out the logical size in the device
  config as well as the vdo config from the test configuration as the
  former is no longer automatically copied from the latter.

.../GrowLogical_t1.c
Added a new test of auto-growth.

.../GrowLogical03.pm
Added a test of auto-growth for unmanaged vdos.

.../DefaultLogicalSize.pm
.../vdotests.suites
Added and disabled a test of auto-growth for lvm managed vdos. This
  test can't reliably pass since this change only addresses the case
  where lvm rounds up.

.../VolumeGroup.pm
Modified createVDOVolume() to not specify a logical size if the
  logicalSize parameter is 0 in order to facilitate testing of the
  default logical size with lvm managed vdos.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156767]