Fixed hang introduced by 156803 [VDO-5467] [BZ 2125708].
While working on IOError_t1 (the unit test added for 156803), it became
apparent that an I/O error on the data write from a hash lock agent
would end up doing the write again, even though the VDO was going to go
read-only since the data_vio already had an error and the user bio was
already acknowledged. It also appeared to be the case that this extra
write caused an infinite loop in the unit test. The unit test loop was
fixed by modifying start_bypassing() to set an error handler so that
agents with errors wouldn't attempt to write again. Unfortunately, this
inadvertantly resulted in a different clean up path for the agent's
hash lock, which, if there were other waiters on the lock, would be
left hanging.

The original unit test hang appears to have been some other test issue
because I can no longer reproduce it. Fixing the extra write will fall
out naturally from the next set of error-path clean ups, so this change
merely reverts the behavior of start_bypassing.

Pair: sclafani

.../dedupe.c
Removed setting of the agent's error handler at the end of
  start_bypassing().

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156810]