Shortened CacheThrashingPerf run time by writing less data [VDO-5421]
Work is in progress to understand why CacheThrashingPerf has
gotten so much slower on linux-vdo, but in the meantime the issue
of a ~3 hour perf test in nightly remains, and even a 90-minute
test is not ideal.

To force VDO page cache evictions, the test writes randomly to a
210GB address space while having a block map only capable of
caching pages addressing 101.5GB of logical space. By default, fio
writes to "fill" the logical space by writing 210GB of data. The
test can still accurately measure performance while writing a much
smaller dataset since the cost of writing a block map page is
independent of how full it is. Fortunately fio has an io_size
parameter that allows the data volume to be specified separately
from the logical space. Limiting it to 40GB reduces the test run
time from 160 to 30 minutes while still reporting the same
performance.

Pair: msakai

.../FIO.pm
Added a ioSize property that, if defined, is passed to fio as the
  io_size command-line parameter.

.../VDOFIOPerfTestBase.pm
Added an ioSize test property (which is automatically a test
  command-line parameter).
Added ioSize to the list of properties run through parseBytes() so
  human-friendly strings like "1G" will become a numeric value.
Passed the ioSize test property as the new FIO command property.
Re-sorted the statements setting FIO command properties.

.../FIOUtils.pm
Updated the assertion that fio wrote as much data as we expected
  to use the smaller of ioSize and total fio job size.

.../vdotests.suites
Made CacheThrashingPerf specify --ioSize=40GB.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156812]