Moved lock-counter.c into recovery-journal.c [VDOUP-6].
As part of the task to consolidate source files, this change moves
the entirety of lock-counter.c into recovery-journal.c, which is
the only file that calls any of the functions in lock-counter.h.
This change is primarily code motion for ease of review. Cleaning up
and unexposing lock counter functions will follow in the next change.

Pair: sclafani

.../lock-counter.c
.../recovery-journal.c
.../Makefile.common
.../MANIFEST.yaml
Moved the contents of lock-counter.c into recovery-journal.c and
  deleted the former.

.../recovery-journal.c
Renamed assert_on_journal_thread() from lock-counter.c to
  assert_lock_counter_on_journal_thread() to avoid a name conflict with
  the already existing function of the same name. This will be cleaned
  up in the next change.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156870]