Do a vdo_dump_all() when timing out a unit test.
Prior to this change, vdo_dump_all() was only compiled for the kernel
build. Compiling it for the user space build was previously deferred as
doing required adding many ifdefs or cases due to the different meaning
of %llu in the kernel vs user space. This change finally addresses
these issues.

Pair: raeburn

.../logger.h
Removed the __printf() attribute from uds_log_message() for internal
  builds, thereby preventing format typecheck errors when attempting to
  compile dump.c in the unit test build.

.../logger.c
Added pragmas around uds_log_message() to turn off the
  missing-format-attribute warning.

.../dump.c
Converted loop variables to be unsigned in parse_dump_options() as
  building in user space requires (it does no harm in the kernel).
Made the computation of DIGITS_PER_UINT64_T into something that the
  user space build believes is a constant.
Added missing include of work-queue.h.

.../data-vio-pool.c
.../data-vio.c
Removed __KERNEL__ ifdefs related to dumping.

.../dump.h
Created a unit-test version of dump.h which wraps the one in the base
  directory, add adds a stub implementation of
  report_uds_memory_usage() which is called from dump.c.

.../workQueue.c
Added unit-test implementations of dump_completion_to_buffer() and
  dump_work_queue() as these are used by dump.c, but the kernel
  implementations in work-queue.c are neither appropriate nor
  accessible to unit tests.

.../Makefile
Added dump.o to TEST_COMMON and a rule to copy dump.c from the base
  directory into the test directory with %llu converted to %lu. Along
  with the test version of dump.h this gives unit tests access to
  vdo_dump_all().

.../vdotest.c
Moved signal masking from the no-fork arm of runTest() into testSub()
  so that signals are correctly blocked whether forking or not. 
Converted use of sigprocmask() to pthread_sigmask() as the man pages
  state is correct for multi-threaded programs.
Modified testSub() to not only set an alarm if a timeout is specified,
  but to also set a signal handler for SIGALRM which will call
  vdo_dump_all() in addition to exiting if the alarm fires.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156899]