Use a user space utility to fill the index for tests.
The storage underlying a VDO volume is not readily available
to user space programs whether the VDO is active or not, but
it can be made accessible by using dmsetup to map a logical
volume over the storage. This method is used by other user
space programs that modify the persisted state of the VDO.

The UDS index can then be opened by a user space program and
filled by feeding it synthetic requests until it is full and
starts discarding its oldest records. It is not important
exactly how many added records are required to cause entries
to be discarded. The program checks the stats approximately
every chapter.

Pair: raeburn, sclafani

.../vdoFillIndex.c
Added user space utility to fill an index.

.../Makefile
Added new vdoFillIndex to the list of program targets.
Made the library dependencies explicit so that a make in the
  vdo/user directory will make the libraries if necessary.

.../BlockDevice/VDO.pm
Added methods to dump the VDO config and use the dumped config
  parameters to open and fill the index with the new
  vdoFillIndex utility.

.../VDOTest.pm
Added the new utility to the list of files to be provided to
  the test machine.

.../BlockDevice03.pm
Revised method for verifying that the index was filled: the
  number of entries indexed should exceed the capacty of a
  full volume and adding another chapter should cause the
  oldest entries to be discarded.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156948]