Add gdb "vdo-dump" command for debugging tests.
The new gdb command is
    vdo-dump <expr> <word>...
where the pointer-to-vdo expression must not contain
whitespace, and the "word" list contains strings like
"viopool" or "queues", etc., just like the "dmsetup
message...dump..." command takes. The arguments are separated
at whitespace, hence the requirement that the VDO pointer
argument not contain whitespace.

The Python file can be automatically loaded by GDB when it
examines the library for debug info. Unfortunately, due to
security factors in modern GDB (akin to shell configurations
not including "." in $PATH by default any more), a little
extra work (command-line arguments or personal config settings
or manually running "source") is needed to tell it that
scripts in this directory are okay to load. If such measures
aren't used, GDB will warn about this including the commands
to change the settings (which means the GDB session is a
little more noisy with this file present but unloaded); the
changes are also mentioned in the file comments.

Pair: sclafani

.../libvdoTest.so-gdb.py
New file.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 156978]