Skip error code setting for data_vio continuation on success.
This is a small clean up which avoids attempting to set an error code
when continuing a data_vio if the code is known to be success. This
makes data_vio continuation a bit clearer as well.

Pair: msakai

.../data-vio.h
Renamed continue_data_vio() to continue_data_vio_with_error().
Added new version of continue_data_vio() which does not take a result
  parameter to replace previous usages of
  continue_data_vio(data_vio, VDO_SUCCESS).

.../data-vio.c
.../block-map.c
.../dedupe.c
.../packer.c
.../physical-zone.c
.../recovery-journal-block.c
.../recovery-journal.c
.../slab-journal.c
Reflected changes.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157009]