Address checkpatch.pl warnings suggesting addition of const.
checkpatch.pl suggests, via warnings, the addition of const to various
data structures.  Generally this changeset adds such to address the
warnings.

In the case of work-queue.c this doesn't work as it results in the
reintroduction of the problem (gcc format check error) that the data
structure (a non-const array) was used to avoid.  For work-queue.c the
changeset drops the non-const array and uses a pragma to suppress the
format check error.

Pair: jwiele

.../admin-completion.h
.../admin-completion.c
.../completion.c
.../data-vio.c
.../dedupe.c
.../pool-sysfs.c
.../vdo-load.c
.../vdo-page-cache.c
.../vdo-resize-logical.c
.../vdo-resize.c
.../vdo-resume.c
.../vdo-suspend.c
Added const to various data structures to silence checkpatch.pl
  warnings.

.../work-queue.c
Replaced use of non-const array with pragma suppressing the format
  check error.  This resolves the checkpatch.pl warning.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157082]