VDOUP-6 Eliminate #ifdefs from upstream code.
Note that there are still two files in the resulting code that
contain #ifdefs unrelated to kernel version: __BYTE_ORDER__ in
murmurhash3 and MODULE in vdo.c.

MODULE is set by the kernel Makefile when a driver is built as
a loadable driver module as opposed to a built in driver. DM
targets, including dm-vdo, are generally defined as tristate
modules, the three states being not to be built at all, to be
built into the kernel image, or to be built as loadable
modules. If MODULE is defined, THIS_MODULE is a pointer to a
module structure that contains, among other things, the module
name as used by vdo.c. If MODULE is not defined, THIS_MODULE
is NULL. A few other drivers use #ifdef MODULE in a similar
way.

__BYTE_ORDER__ is necessary to maintain portability between
hardware platforms. It comes from the upstream murmurhash3
implemention. There may be a preferred way to handle this.

A follow on changes will factor out common stuff from
packaging/github and packaging/kpatch. Another change will add
target(s) and variables to packaging/kpatch/Makefile for
feeding the script that mirrors the code to the public
repository so that the script can be simplified.

Pair: msakai, jshimkus, corwin

.../MANIFEST.yaml
Added kpatch MANIFEST to do special ifdef processing for code
  intended for the upstream kernel.

.../Makefile
Fixed to use the kpatch specific manifest and provide the
  check-style target.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 157218]